Tarefa1_2021li1g033

Description

O objetivo desta tarefa é implementar uma função que dada uma lista de peças e respetivas coordenadas testa se estas definem corretamente um mapa. Para tal, a lista deverá satisfazer os seguintes critérios (1.1 - 1.5).

validaPotencialMapa

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Verifica se o mapa é válido.

Exemplos de utilização:

>>> validaPotencialMapa [(Porta,(0,3)),(Bloco,(0,4)),(Bloco,(1,4)),(Bloco,(2,4)),(Bloco,(3,4)),(Bloco,(4,4)),(Caixa,(4,3)),(Bloco,(5,4)),(Bloco,(6,4)),(Bloco,(6,3)),(Bloco,(6,2)),(Bloco,(6,1))]
True
>>> validaPotencialMapa []
False

1.1 - Não haver mais do que uma declaração de peça para a mesma posição.

coordenadaDiferente

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Verifica se as peças têm todas coordenadas diferentes.

Exemplos de utilização:

>>> coordenadaDiferente [(Bloco, (0,2)),(Bloco,(0,3))]
True
>>> coordenadaDiferente [(Caixa,(0,1)),(Bloco,(0,1))]
False

pertenceCauda

Arguments

:: Coordenadas
Uma lista de pares de Int.

-> [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Auxiliar de coordenadaDiferente - Verifica se uma coordenada não pertence à cauda (tail).

Exemplos de utilização:

>>> pertenceCauda (1,2) [(Bloco,(0,3)),(Caixa,(1,2))]
True
>>> pertenceCauda (0,0) [(Bloco,(3,2)),(Caixa,(1,1))]
False

coordenadaPositiva

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Verifica se as coordenadas são todas positivas, visto que o mapa tem de começar nas coordenadas (0,0).

Exemplos de utilização.

>>> coordenadaPositiva [(Bloco,(0,2)),(Bloco,(0,4))]
True
>>> coordenadaPositiva [(Caixa,(-1,2)),(Bloco,(2,-1))]
False
>>> coordenadaPositiva [(Porta,(-1,-2)),(Bloco,(2,1))]
False

1.2 - Declarar exatamente uma porta.

portaUnica

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Verifica se existe uma e uma só porta.

Exemplos de utilização:

>>> portaUnica [(Porta,(1,2)),(Bloco,(1,3)),(Bloco,(1,4))]
True
>>> portaUnica [(Porta,(1,2)),(Porta,(1,3)),(Caixa,(2,4))]
False

contaPortas

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Int
Resultado.

Auxiliar de portaUnica - Conta quantas portas existem no mapa.

1.3 - Todas as caixas devem estar posicionadas em cima de outra caixa ou bloco, i.e. não podem haver caixas a "flutuar".

caixasAFlutuar :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)] -> Bool

Verifica se não existem caixas a flutuar.

Exemplos de utilização:

>>> caixasAFlutuarWrapper [(Bloco,(0,1)),(Caixa,(0,0))]
True
>>> caixasAFlutuarWrapper [(Bloco,(0,0)),(Caixa,(1,2))]
False

Observação:

caixasAFlutuarWrapper l = caixasAFlutuar l l -- Onde l é uma lista de de pares de Peca e de Coordenadas.

procuraPeca

Arguments

:: Coordenadas
Um par de Int.

-> [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Auxiliar de caixasAFlutuar - Verifica se existe alguma peça na posição dada (neste caso: y + 1).

1.4 - Devem existir espaços vazios (no mínimo um), i.e. o mapa não pode estar totalmente preenchido por caixas, blocos e porta.

peloMenosUmVazio

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Verifica se existe pelo menos um espaço vazio (declarado ou não) no mapa.

Exemplos de utilização:

>>> peloMenosUmVazio [(Vazio,(1,2)),(Bloco,(3,3))]
True
>>> peloMenosUmVazio [(Bloco,(0,2)),(Bloco,(0,3))] -- Vazio definido por omissão.
True
>>> peloMenosUmVazio [(Bloco,(0,0)),(Bloco(1,0))]
False

existemDeclarados

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Auxiliar de peloMenosUmVazio - Conta quantos vazios declarados existem.

existemNaoDeclarados

Arguments

:: Int
Um Int.

-> [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Bool
Resultado.

Auxiliar de peloMenosUmVazio - Conta quantos vazios não declarados existem.