Tarefa2_2021li1g033

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Construção do Mapa.

constroiMapa

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Mapa
Um Mapa.

Transforma uma lista de pares de Peca e de Coordenadas num Mapa.

Exemplos de utilização:

>>> constroiMapa [[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Vazio],[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],[Vazio, Vazio, Vazio, Vazio, Vazio, Vazio, Bloco],[Porta, Vazio, Vazio, Vazio, Caixa, Vazio, Bloco],[Bloco, Bloco, Bloco, Bloco, Bloco, Bloco, Bloco]]
[(Porta, (0, 3)),(Bloco, (0, 4)),(Bloco, (1, 4)),(Bloco, (2, 4)),(Bloco, (3, 4)),(Bloco, (4, 4)),(Caixa, (4, 3)),(Bloco, (5, 4)),(Bloco, (6, 4)),(Bloco, (6, 3)),(Bloco, (6, 2)),(Bloco, (6, 1))]

constroiMapaVazio

Arguments

:: [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

-> Mapa
Um Mapa.

Constrói uma lista com um número de listas vazias igual à altura do mapa (y), cujo cada lista contém Vazio (até à largura (x) máxima).

inserePeca

Arguments

:: (Peca, Coordenadas)
Um par de Peca e de Coordenadas.

-> Mapa
Um Mapa.

-> Mapa
Um Mapa.

Insere cada uma das peças no espaço correspondente.

Desconstrução do Mapa.

desconstroiMapa

Arguments

:: Mapa
Um Mapa.

-> [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

Função inversa de constroiMapa.

desconstroiPeca

Arguments

:: Mapa
Um Mapa.

-> Int
Um Int - neste caso serve como acumulador.

-> [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

Desconstrói as peças.

desconstroiLinha

Arguments

:: [Peca]
Uma lista de Peca - lista a desconstruir.

-> Int
Um Int - abcissa da Peca.

-> Int
Um Int - ordenada da Peca.

-> [(Peca, Coordenadas)]
Uma lista de pares de Peca e de Coordenadas.

Auxiliar de descontroiPeca - Desconstrói as peças por linha.