Tarefa4_2021li1g033

Description

O objetivo desta tarefa é implementar a função moveJogador que aplica o efeito de um comando (i.e. Movimento) sobre o jogador, e a sua generalização correrMovimentos que aplica consecutivamente os comandos dados pela lista.

Documentation

moveJogador

Arguments

:: Jogo
Um Jogo.

-> Movimento
Um Movimento a aplicar.

-> Jogo
Resultado - O Jogo depois de se ter aplicado o Movimento.

Verifica se o Movimento é aplicável.

correrMovimentos

Arguments

:: Jogo
Um Jogo.

-> [Movimento]
Uma lista de Movimento.

-> Jogo
Resultado - O Jogo depois de lhe ter sido aplicada a lista de Movimento.

Aplica cada uma lista de Movimento ao Jogo.

moveEsquerda

Arguments

:: Jogo
Um Jogo.

-> Jogo
Resultado.

Aplica o Movimento: AndarEsquerda.

moveDireita

Arguments

:: Jogo
Um Jogo.

-> Jogo
Resultado.

Aplica o Movimento: AndarDireita.

podeTrepar

Arguments

:: Jogo
Um Jogo.

-> Jogo
Resultado.

Aplica o Movimento: Trepar.

interageCaixa

Arguments

:: Jogo
Um Jogo.

-> Jogo
Resultado.

Aplica o Movimento: InterageCaixa.