Utils

Documentation

alturaMapa :: [(Peca, Coordenadas)] -> Int

Calcula a altura de um mapa.

larguraMapa :: [(Peca, Coordenadas)] -> Int

Calcula a largura de um mapa.

pecasNaLinha :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Encontra todas as peças declaradas numa linha.

ordenaMapa :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Ordena um mapa por ordem crescente das suas coordenadas.

listaCoordenadas :: [(Peca, Coordenadas)] -> [Coordenadas]

Transforma a lista dada, numa lista apenas com as segundas componentes (neste caso coordenadas).