Utils

Documentation

alturaMapa :: [(Peca, Coordenadas)] -> Int

Calcula a altura de um Mapa.

larguraMapa :: [(Peca, Coordenadas)] -> Int

Calcula a largura de um Mapa.

pecasNaLinha :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Calcula a altura de um Mapa.

Encontra todas as peças declaradas numa linha.

ordenaMapa :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Ordena um mapa por ordem crescente das suas Coordenadas.

listaCoordenadas :: [(Peca, Coordenadas)] -> [Coordenadas]

Transforma a lista dada numa lista apenas com as segundas componentes (neste caso Coordenadas).

menorOrdenadaWrapper :: [(Peca, Coordenadas)] -> Int

menorOrdenada :: [Coordenadas] -> Int

Calcula a menor ordenada de um Mapa.

retiraPeca :: Coordenadas -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Retira a Peca que se encontra nas Coordenadas dadas.

pecasNaColuna :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Encontra todas as peças declaradas numa coluna.

extraiBlocos :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Extrai apenas os Bloco que existem na lista.

retiraColuna :: Int -> [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Retira a coluna dada à lista.

coordenadaPertence :: Coordenadas -> [Coordenadas] -> Bool

Verifica se um par de Coordenadas pertence à lista dada.

encontraPeca :: Coordenadas -> Mapa -> Maybe Peca

Encontra a Peca que se encontra numas determinadas Coordenadas.

encontraNaLinha :: Int -> [Peca] -> Maybe Peca

Auxiliar de encontraNaLinha - Procura pela Peca na linha.

pecaEsperada :: Peca -> Coordenadas -> Mapa -> Bool

pecaImpossivel :: Coordenadas -> Mapa -> Bool

encontraBaixo :: Coordenadas -> Mapa -> Coordenadas

substituiPeca :: Coordenadas -> Peca -> Mapa -> Mapa

substituiPecaWrapper :: Coordenadas -> Peca -> Mapa -> Mapa -> Mapa

substituiPecaNaLinha :: Int -> Peca -> [Peca] -> [Peca]

substituiPecaNaLinhaWrapper :: Int -> Peca -> [Peca] -> [Peca] -> [Peca]